requirements.txt
pandas
requests
spacetrack
numpy
matplotlib
scikit-learn
torch
python-dotenv
sgp4
